move to tax_lot_properties
